Add eval utilities module and refactor imports
