+++ econ499/eval/eval_walk_forward.py
@@ -8,12 +8,8 @@ import pandas as pd
 
 from iv_drl.utils import load_config
 from iv_drl.utils.metrics_utils import rmse, mae
-from iv_drl.evaluation.evaluate_all import (
-    _load_predictions,
-    _find_forecast_col,
-    _mape,
-    _qlike,
-)
+from econ499.eval.utils import _load_predictions, _mape, _qlike
+from econ499.eval.evaluate_all import _find_forecast_col
 
 CFG = load_config("data_config.yaml")
 ROOT = Path(__file__).resolve().parents[2]
