+++ econ499/eval/evaluate_all.py
@@ -13,6 +13,7 @@ import pandas as pd
 
 from econ499.utils import load_config
 from econ499.utils.metrics_utils import rmse, mae
+from econ499.eval.utils import _load_predictions, _mape, _qlike
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 CONFIG = load_config("data_config.yaml")
@@ -20,11 +21,6 @@ ROOT = Path(__file__).resolve().parents[2]
 DATA_DIR = Path(CONFIG["paths"]["output_dir"]).resolve()
 ARTIFACT_DIR = ROOT / "artifacts" / "tables"
 TARGET_COL = CONFIG["features"]["target_col"]
-
-
-def _mape(y: np.ndarray, yhat: np.ndarray) -> float:
-    mask = y != 0
-    return float(np.mean(np.abs((y[mask] - yhat[mask]) / y[mask])) * 100)
 
 
 def _find_forecast_col(df: pd.DataFrame) -> str:
@@ -37,36 +33,6 @@ def _find_forecast_col(df: pd.DataFrame) -> str:
     raise ValueError("Unable to identify forecast column")
 
 
-def _load_predictions() -> list[tuple[str, pd.Series]]:
-    out = []
-    for csv in DATA_DIR.glob("*_oos_predictions.csv"):
-        try:
-            df = pd.read_csv(csv, parse_dates=["date"])
-        except Exception:
-            continue
-        if df.empty or "date" not in df.columns:
-            continue
-        col = _find_forecast_col(df)
-        name = col.replace("_forecast", "")
-        out.append((name, df[["date", col]].rename(columns={col: name})))
-    return out
-
-
-def _qlike(y: np.ndarray, yhat: np.ndarray, eps: float = 1e-12) -> float:
-    """QLIKE loss (Patton & Sheppard, 2009) for volatility forecasts.
-
-    Parameters
-    ----------
-    y : np.ndarray
-        Realised volatility proxy (here next-day ATM-IV).
-    yhat : np.ndarray
-        Forecasts (same scale as *y*).
-    eps : float, optional
-        Floor to avoid log/zero division when predictions are ~0.
-    """
-    y = np.asarray(y, dtype=float)
-    yhat = np.asarray(yhat, dtype=float).clip(min=eps)
-    return float(np.mean(np.log(yhat) + y / yhat))
 
 
 def evaluate_all(panel_csv: str | Path | None = None, **kwargs) -> Path:
